[Avinash] : Added type casting in the program for calculating profit and loss

+++ gcr-codebase/programming-elements/Level-1/CalculateProfitLoss.java
@@ -3,7 +3,7 @@ public class CalculateProfitLoss {
 		int costPrice = 129, sellingPrice = 191;
 		
 		int profit = sellingPrice - costPrice;
-		double profitPercent = (double)profit / (costPrice * 100.0);
+		double profitPercent = (double)profit / (double)costPrice * 100.0;
 		
 		System.out.println("The Cost Price is INR " + costPrice + " and Selling Price is INR " + sellingPrice);
 		System.out.println("The Profit is INR " + profit + " and the Profit Percentage is " + profitPercent);
